moves variables to preflight plugin
