+++ package.json
@@ -36,7 +36,9 @@
     "svelte-preprocess": "^4.0.0",
     "ts-jest": "^27.0.3",
     "tslib": "^2.0.0",
-    "typescript": "^4.0.0"
+    "typescript": "^4.0.0",
+    "postcss": "^8.4.14",
+    "postcss-js": "^4.0.0"
   },
   "scripts": {
     "install": "npm run transform-tokens",
