+++ tokens/transformation/tailwind/static/index.js
@@ -58,4 +58,7 @@ module.exports = {
 			},
 		},
 	},
+	plugins: [
+		require("./preflightPlugin")
+	],
 };
