+++ tokens/transformation/tailwind/static/preflightPlugin.js
@@ -1,0 +1,8 @@
+const plugin = require("tailwindcss/plugin");
+const variables = require("./_variables.css.js");
+
+module.exports = plugin(function ({ addBase }) {
+	addBase({
+        ...variables
+    });
+});
